separate cluster and application tests
