+++ azure-cli2017.pyproj
@@ -843,7 +843,9 @@
     <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\commands.py" />
     <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\custom.py" />
     <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\operations\applications.py" />
-    <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\tests\latest\test_sf_commands.py" />
+    <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\tests\latest\test_util.py" />
+    <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\tests\latest\test_sf_application.py" />
+    <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\tests\latest\test_sf_cluster.py" />
     <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\_arm_deployment_utils.py" />
     <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\_client_factory.py" />
     <Compile Include="azure-cli\azure\cli\command_modules\servicefabric\_help.py" />
